remove twig_component tag and move slot logic into component tag
