+++ src/TwigComponent/src/Twig/ComponentExtension.php
@@ -49,8 +49,7 @@ final class ComponentExtension extends AbstractExtension implements ServiceSubsc
     public function getTokenParsers(): array
     {
         return [
-            new ComponentTokenParser(fn () => $this->container->get(ComponentFactory::class)),
-            new TwigComponentTokenParser(fn () => $this->container->get(ComponentFactory::class), $this->environment),
+            new ComponentTokenParser(fn () => $this->container->get(ComponentFactory::class), $this->environment),
             new SlotTokenParser(),
         ];
     }
